+++ app/Http/Controllers/StatusController.php
@@ -254,6 +254,7 @@ class StatusController extends Controller
 			$share->profile_id = $profile->id;
 			$share->reblog_of_id = $status->id;
 			$share->in_reply_to_profile_id = $status->profile_id;
+			$share->type = 'share';
 			$share->save();
 			$count++;
 			SharePipeline::dispatch($share);